Support 64-bit loop indices on GPU backends and fix scan memory waste
GPU codegen now uses 64-bit integers for loop variables and indices
(long on Metal, long long on CUDA/HIP), enabling grids with more than
2^31 elements. Dispatch passes loop_end as c_longlong. Type inference
auto-promotes large Python int args to i64.

Fix exclusive/inclusive scan copying entire multi-GB buffer to numpy
just to read the last element — now uses a single-element kernel read.

Add flying edges (example 27) and marching cubes (example 26) examples.

+++ src/pgc/runtime/cuda_backend.py
@@ -208,7 +208,7 @@ class CompiledCUDAKernel:
 
     def __call__(self, kernel_args: list, loop_end: int):
         """Dispatch the CUDA kernel."""
-        n_val = ctypes.c_int(loop_end)
+        n_val = ctypes.c_longlong(loop_end)
 
         arg_values = []
         for arg, ptype, is_field in zip(kernel_args, self._param_types, self._param_is_field):
